requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.26.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
databases==0.6.1
ecdsa==0.18.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.3.0
mypy-extensions==1.0.0
orjson==3.9.0
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
